# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SE Project API (Team 8, May 2025)
  description: "Backend API for Team 8's Software Engineering Project (May 2025 Term)\n\
    \nThis backend powers a web-based mobile application designed for senior citizens\
    \ and their family caregivers. The application enhances daily living and care\
    \ coordination through these core modules identified based on user stories:\n\n\
    - Easy-to-use UI: Senior-friendly design with large text, bold colors, and simple\
    \ navigation\n- Medicine Tracker: Prescription management and medication identification\n\
    - Appointments Tracker: Book, view, and get reminders for doctor appointments\n\
    - SOS: Emergency alert system for designated contacts\n- Documents Vault: Secure\
    \ digital repository for medical reports and history\n\n---\n\n## User Story Mapping\n\
    \n| # | User Story (Summary) | API Endpoints |\n|---|----------------------|--------------|\n\
    | 1 | UI accessibility | (Frontend/UI, not API) |\n| 2 | Medicine management |\
    \ `/medicines/`, `/medicines/{medicine_id}`, `/medicines/transcribe` |\n| 3 |\
    \ Appointment management | `/appointments/`, `/appointments/user/{user_id}` |\n\
    | 4 | Emergency contact notification | `/users/{user_id}/sos/trigger`, `/emergency-contacts/`\
    \ |\n| 5 | Document storage | `/documents/`, `/documents/upload`, `/documents/user/{user_id}`\
    \ |\n| 6 | Medication adherence tracking | `/medicines/`, `/medicines/user/{user_id}`\
...
paths:
...
  /api/v1/auth/sms/send:
    post:
      tags:
      - Authentication
      summary: Send SMS verification code
      description: Send a verification code to the provided phone number. This is
        step 1 of the SMS verification process.
      operationId: send_sms_verification_api_v1_auth_sms_send_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSVerificationRequest'
        required: true
      responses:
        '200':
          description: Verification code sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSVerificationResponse'
        '400':
          description: Invalid phone number format
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to send verification code
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/sms/verify:
    post:
      tags:
      - Authentication
      summary: Verify SMS code
      description: Verify the SMS code received on the phone number. This is step
        2 of the SMS verification process.
      operationId: verify_sms_code_api_v1_auth_sms_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSVerificationCodeRequest'
        required: true
      responses:
        '200':
          description: Code verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSVerificationStatusResponse'
        '400':
          description: Invalid code or phone number
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to verify code
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/sms/status/{phone}:
    get:
      tags:
      - Authentication
      summary: Get SMS verification status
      description: Check the current verification status for a phone number.
      operationId: get_sms_verification_status_api_v1_auth_sms_status__phone__get
      parameters:
      - name: phone
        in: path
        required: true
        schema:
          type: string
          description: Phone number to check verification status for
          title: Phone
        description: Phone number to check verification status for
      responses:
        '200':
          description: Verification status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSVerificationStatusResponse'
        '400':
          description: Bad Request - Invalid input data
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to get verification status
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/passkey/register/challenge:
    post:
      tags:
      - Authentication
      summary: Create passkey registration challenge
      description: Generate a WebAuthn registration challenge for passkey setup. This
        is step 1 of passkey registration.
      operationId: create_passkey_registration_challenge_api_v1_auth_passkey_register_challenge_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyRegistrationRequest'
        required: true
      responses:
        '200':
          description: Registration challenge created successfully
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Passkey Registration Challenge Api V1 Auth
                  Passkey Register Challenge Post
        '400':
          description: Invalid user data
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to create challenge
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/passkey/register/verify:
    post:
      tags:
      - Authentication
      summary: Verify passkey registration
      description: Verify WebAuthn registration response and create passkey credential.
        This is step 2 of passkey registration. Sets session cookie on success.
      operationId: verify_passkey_registration_api_v1_auth_passkey_register_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_verify_passkey_registration_api_v1_auth_passkey_register_verify_post'
        required: true
      responses:
        '200':
          description: Registration verified successfully, user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyVerificationResult'
        '400':
          description: Invalid registration response
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to verify registration
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/passkey/login/challenge:
    post:
      tags:
      - Authentication
      summary: Create passkey login challenge
      description: Generate a WebAuthn authentication challenge for passkey login.
        This is step 1 of passkey authentication.
      operationId: create_passkey_login_challenge_api_v1_auth_passkey_login_challenge_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasskeyLoginRequest'
        required: true
      responses:
        '200':
          description: Login challenge created successfully
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Passkey Login Challenge Api V1 Auth Passkey
                  Login Challenge Post
        '400':
          description: Invalid credential ID
        '401':
          description: Unauthorized - Authentication required
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to create challenge
        '404':
          description: Credential not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/passkey/login/verify:
    post:
      tags:
      - Authentication
      summary: Verify passkey login
      description: Verify WebAuthn authentication response for passkey login. This
        is step 2 of passkey authentication. Sets session cookie on success.
      operationId: verify_passkey_login_api_v1_auth_passkey_login_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_verify_passkey_login_api_v1_auth_passkey_login_verify_post'
        required: true
      responses:
        '200':
          description: Login verified successfully, session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasskeyVerificationResult'
        '400':
          description: Invalid login response
        '401':
          description: Authentication failed
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Failed to verify login
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/passkey/user/{user_id}:
    get:
      tags:
      - Authentication
      summary: Get user passkeys
      description: Retrieve all passkey credentials for a specific user. Requires
        authentication and ownership.
      operationId: get_user_passkeys_api_v1_auth_passkey_user__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: List of user's passkey credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PasskeyCredentialResponse'
                title: Response Get User Passkeys Api V1 Auth Passkey User  User Id  Get
        '400':
          description: Bad Request - Invalid input data
        '401':
          description: Not authenticated
        '403':
          description: Not authorized to access this user's data
        '500':
          description: Internal Server Error
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: User logout
      description: Logout user by invalidating session and clearing session cookie.
      operationId: logout_user_api_v1_auth_logout_post
      parameters:
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema: {}
              example:
                message: Logout successful
        '400':
          description: Failed to logout
        '401':
          description: Not authenticated
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Internal Server Error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user info
      description: Get information about the currently authenticated user.
      operationId: get_current_user_info_api_v1_auth_me_get
      parameters:
      - name: session_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Session Token
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request - Invalid input data
        '401':
          description: Not authenticated
        '403':
          description: Forbidden - Insufficient permissions
        '500':
          description: Internal Server Error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_verify_passkey_login_api_v1_auth_passkey_login_verify_post:
      properties:
        request:
          $ref: '#/components/schemas/PasskeyLoginRequest'
        response_data:
          additionalProperties: true
          type: object
          title: Response Data
      type: object
      required:
      - request
      - response_data
      title: Body_verify_passkey_login_api_v1_auth_passkey_login_verify_post
    Body_verify_passkey_registration_api_v1_auth_passkey_register_verify_post:
      properties:
        request:
          $ref: '#/components/schemas/PasskeyRegistrationRequest'
        response_data:
          additionalProperties: true
          type: object
          title: Response Data
      type: object
      required:
      - request
      - response_data
      title: Body_verify_passkey_registration_api_v1_auth_passkey_register_verify_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PasskeyLoginRequest:
      properties:
        credential_id:
          type: string
          title: Credential Id
          description: Credential ID for login (optional)
          example: cred-123
      type: object
      title: PasskeyLoginRequest
      description: Request to start passkey login.
    PasskeyRegistrationRequest:
      properties:
        user_phone:
          type: string
          title: User Phone
          description: User phone number for registration
          example: '+919876543210'
        user_name:
          type: string
          title: User Name
          description: User name for registration
          example: Amit Sharma
        user_dob:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: User Dob
          description: User date of birth (optional)
          example: '1950-01-01'
        user_gender:
          anyOf:
          - type: string
          - type: 'null'
          title: User Gender
          description: User gender (optional)
          example: Male
      type: object
      required:
      - user_phone
      - user_name
      title: PasskeyRegistrationRequest
      description: Request to start passkey registration.
    PasskeyVerificationResult:
      properties:
        user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: User Id
          description: User ID if login was successful (optional)
          example: 1
        credential_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Credential Id
          description: Credential ID used (optional)
          example: cred-123
        session_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Session Expires At
          description: Session expiration time if applicable (optional)
          example: '2025-07-01T12:00:00Z'
      type: object
      title: PasskeyVerificationResult
      description: Result of passkey verification.
    SMSVerificationCodeRequest:
      properties:
        phone:
          type: string
          title: Phone
          description: Phone number that received the verification code (E.164 format)
          example: '+1234567890'
        code:
          type: string
          maxLength: 8
          minLength: 4
          title: Code
          description: Verification code received via SMS (4-8 digits)
          example: '123456'
      type: object
      required:
      - phone
      - code
      title: SMSVerificationCodeRequest
      description: Schema for SMS verification code submission. Used to verify the
        code received via SMS.
      example:
        code: '123456'
        phone: '+1234567890'
    SMSVerificationRequest:
      properties:
        phone:
          type: string
          title: Phone
          description: Phone number to send verification code to (E.164 format)
          example: '+1234567890'
      type: object
      required:
      - phone
      title: SMSVerificationRequest
      description: Schema for SMS verification initiation request. Used to request
        a verification code via SMS.
      example:
        phone: '+1234567890'
    SMSVerificationResponse:
      properties:
        success:
          type: boolean
          title: Success
          description: Whether the operation was successful
          example: true
        message:
          type: string
          title: Message
          description: Response message
          example: Verification code sent successfully
        expires_in:
          anyOf:
          - type: integer
          - type: 'null'
          title: Expires In
          description: Time in seconds until verification expires (optional)
          example: 600
      type: object
      required:
      - success
      - message
      title: SMSVerificationResponse
      description: Schema for SMS verification response. Indicates if the code was
        sent and when it expires.
      example:
        expires_in: 600
        message: Verification code sent successfully
        success: true
    SMSVerificationStatusResponse:
      properties:
        verified:
          type: boolean
          title: Verified
          description: Whether the phone number is verified
          example: true
        message:
          type: string
          title: Message
          description: Response message
          example: Phone number successfully verified
        expires_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Expires At
          description: When the verification status expires (ISO 8601, optional)
          example: '2025-06-28T22:30:00Z'
      type: object
      required:
      - verified
      - message
      title: SMSVerificationStatusResponse
      description: Schema for SMS verification status response. Indicates if the phone
        number is verified and expiry info.
      example:
        expires_at: '2025-06-28T22:30:00Z'
        message: Phone number successfully verified
        verified: true
...